Made some modifications in styling of hackathons page
