+++ devagram/src/Components/JobsAndHacks/HackathonCard/Details/Details.module.css
@@ -24,6 +24,7 @@
   left: 50%;
   transform: translate(-50%, -50%);
   background-color: white;
+  border-radius: 20px;
 }
 
 .event_logo {
