+++ devagram/src/Components/JobsAndHacks/HackathonCard/HackathonCard.module.css
@@ -4,10 +4,19 @@
   flex-direction: column;
   align-items: center;
   overflow: hidden;
-  border: 3px solid black;
+  animation: scl 0.5s ease-in-out;
+  transform-origin: left center;
+  /* border: 3px solid black; */
+  border-radius: 20px;
+  box-shadow: 2px 2px 25px 0px rgba(0, 0, 0, 0.4);
+}
+.hackathoncard:hover {
+  box-shadow: 7px 7px 15px 2px rgba(0, 0, 0, 0.5);
+  transition: box-shadow 0.3s ease-in;
 }
 .hackathoncard img {
-  object-fit: contain;
+  /* object-fit: contain; */
+  background-size: cover;
   height: 300px;
   width: 100%;
 }
@@ -37,6 +46,13 @@
   border: none;
   outline: 0;
   padding: 6px 15px;
+  border-radius: 4px;
+}
+.hackathoncard__buttons button:hover {
+  background-color: #03429b;
+  box-shadow: 2px 2px 25px 0px rgba(0, 0, 0, 0.4);
+  transition: box-shadow 0.3s ease-in;
+  transition: background-color 0.3s ease-in;
 }
 .learn_more {
   padding: 0em 1em 1em 1em;
